fix engine version stuff
