+++ src/packages/cli/package.json
@@ -64,6 +64,7 @@
   },
   "devDependencies": {
     "@apexearth/copy": "1.4.5",
+    "@jsdevtools/ono": "^7.1.3",
     "@prisma/client": "workspace:*",
     "@prisma/debug": "workspace:*",
     "@prisma/fetch-engine": "workspace:*",
@@ -79,6 +80,7 @@
     "@types/jest": "26.0.14",
     "@types/pg": "7.14.5",
     "@types/sqlite3": "3.1.6",
+    "@types/verror": "^1.10.4",
     "@types/ws": "7.2.6",
     "@typescript-eslint/eslint-plugin": "4.2.0",
     "@typescript-eslint/parser": "4.2.0",
@@ -121,6 +123,7 @@
     "terminal-link": "2.1.1",
     "ts-jest": "26.4.0",
     "typescript": "4.0.3",
+    "verror": "^1.10.0",
     "ws": "7.3.1"
   },
   "scripts": {
