+++ src/packages/cli/src/__tests__/integration/mariadb/runtime.test.ts
@@ -1,6 +1,6 @@
 import * as IntegrationTest from '../../__helpers__/integrationTest'
-import database from './database'
-import scenarios from './scenarios'
+import { database } from './__database'
+import { scenarios } from './__scenarios'
 
 IntegrationTest.runtime({
   database,
